text_to_image/backend/requirements.txt
requests
python-dotenv
Pillow
tqdm
huggingface_hub>=0.26.0

